requirements.txt
blist==1.3.6
defusedxml==0.6.0
Django==2.2.4
django-choices==1.7.0
django-cors-headers==3.1.0
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-extensions==2.2.1
django-filter==2.2.0
djangorestframework==3.10.2
djangorestframework-csv==2.1.0
djangorestframework-xml==1.4.0
drf-amsterdam==0.3.2
drf-extensions==0.5.0
numpy==1.17.0
pandas==0.25.0
psycopg2==2.8.3
pydotplus==2.0.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
raccoon==2.1.5
six==1.12.0
sqlparse==0.3.0
tabulate==0.8.3
unicodecsv==0.14.1
